Refactored code to assign guild.xyz roles to user
